Drop support for deprecated SQLv1
- Restructured tests and removed v1-specific ones
- Updated test factories
- Updated examples
