+++ test/integration/factories/sql_instances_factory.rb
@@ -1,6 +1,6 @@
 require "integration/factories/collection_factory"
 
-class SqlV2InstancesFactory < CollectionFactory
+class SqlInstancesFactory < CollectionFactory
   def initialize(example)
     super(Fog::Google[:sql].instances, example)
   end
@@ -13,7 +13,7 @@ class SqlV2InstancesFactory < CollectionFactory
     # Name has a time suffix due to SQL resources API objects having
     # a _very_ long life on the backend (n(days)) after deletion.
     { :name => "#{resource_name}-#{Time.now.to_i}",
-      :region => TEST_SQL_REGION_SECOND,
-      :tier => TEST_SQL_TIER_SECOND }
+      :region => TEST_SQL_REGION,
+      :tier => TEST_SQL_TIER }
   end
 end
