+++ test/integration/sql/test_common_flags.rb
