+++ test/integration/sql/test_common_tiers.rb
