+++ test/integration/sql/test_coverage.rb
@@ -3,4 +3,4 @@ require "helpers/integration_test_helper"
 # This is a simple coverage helper that helps differentiate
 # the tests when run in parallel so the final coverage report
 # can be properly combined together from multiple runners
-SimpleCov.command_name "test:sqlv1" if ENV["COVERAGE"]
+SimpleCov.command_name "test:sql" if ENV["COVERAGE"]
